utils tests refactor
